requirements.txt
# Core dependencies
flask==3.0.0
edge-tts==6.1.9
numpy==1.26.2
pedalboard==0.9.8

# Audio processing
soundfile==0.12.1

# AI script generation with Google Gemini
google-generativeai==0.3.2

# Language detection
langdetect==1.0.9

# Development
flask-cors==4.0.0
python-dotenv==1.0.0
